Implement InlineView to show documents inline.
We do not have a button for them yet, but documents can now be shown
inline under /view/doc/<number>.

I have subclassed the FetchView rather than adding an additional
parameter to the FetchView as it seems neater to have a separate URL for
inline viewing instead of the same URL but with additional options. It
also makes the intent quite explicit.