Use separate Views for documents and thumbnails.
The intent is to make the code more explicit and more approachable. The
added _DocumentView class simply acts as a base class to retrieve a
document. This then is implemented by specific inheriting classes for
downloads and inline views of documents. The thumbnails are handled
separately as they access a different path and have a hardcoded type and
no disposition.

This way it is clear which function a given View has and which URL
refers to it.